Refactor types, remove unused imports, and improve type safety
