+++ src/components/common/withErrorBoundary.tsx
@@ -1,0 +1,20 @@
+import React, { ReactNode } from 'react';
+import { ErrorBoundary } from './ErrorBoundary';
+
+export const withErrorBoundary = <P extends object>(
+  Component: React.ComponentType<P>,
+  fallback?: ReactNode,
+  onReset?: () => void
+) => {
+  const WrappedComponent = (props: P) => (
+    <ErrorBoundary fallback={fallback} onReset={onReset}>
+      <Component {...props} />
+    </ErrorBoundary>
+  );
+
+  WrappedComponent.displayName = `withErrorBoundary(${
+    Component.displayName || Component.name
+  })`;
+
+  return WrappedComponent;
+};
